refactor: 카드 폼 규칙 상수 합침
