+++ src/components/CardPasswordInputField/constants.ts
@@ -1,4 +1,6 @@
-export const PASSWORD_MAX_LENGTH = 2;
+import { CARD_FORM_LENGTH } from "@/constants/cardForm";
+
+export const PASSWORD_MAX_LENGTH = CARD_FORM_LENGTH.PASSWORD;
 
 export const HELPER_MESSAGE = {
   DEFAULT: "",
