+++ src/components/CardValidityPeriodInputField/constants.ts
@@ -1,5 +1,7 @@
-export const MONTH_MAX_LENGTH = 2;
-export const YEAR_MAX_LENGTH = 2;
+import { CARD_FORM_LENGTH } from "@/constants/cardForm";
+
+export const MONTH_MAX_LENGTH = CARD_FORM_LENGTH.VALIDITY_MONTH;
+export const YEAR_MAX_LENGTH = CARD_FORM_LENGTH.VALIDITY_YEAR;
 
 export const HELPER_MESSAGE = {
   DEFAULT: "",
